fixed pages for menuitemreviews
