--- frontend/src/tests/pages/MenuItemReview/RestaurantIndexPage.test.js
@@ -1,190 +1,0 @@
-import { render, screen, waitFor, fireEvent } from "@testing-library/react";
-import RestaurantIndexPage from "main/pages/Restaurants/RestaurantIndexPage";
-import { QueryClient, QueryClientProvider } from "react-query";
-import { MemoryRouter } from "react-router-dom";
-import mockConsole from "jest-mock-console";
-import { restaurantFixtures } from "fixtures/restaurantFixtures";
-
-import { apiCurrentUserFixtures } from "fixtures/currentUserFixtures";
-import { systemInfoFixtures } from "fixtures/systemInfoFixtures";
-import axios from "axios";
-import AxiosMockAdapter from "axios-mock-adapter";
-
-const mockToast = jest.fn();
-jest.mock("react-toastify", () => {
-  const originalModule = jest.requireActual("react-toastify");
-  return {
-    __esModule: true,
-    ...originalModule,
-    toast: (x) => mockToast(x),
-  };
-});
-
-describe("RestaurantIndexPage tests", () => {
-  const axiosMock = new AxiosMockAdapter(axios);
-
-  const testId = "RestaurantTable";
-
-  const setupUserOnly = () => {
-    axiosMock.reset();
-    axiosMock.resetHistory();
-    axiosMock
-      .onGet("/api/currentUser")
-      .reply(200, apiCurrentUserFixtures.userOnly);
-    axiosMock
-      .onGet("/api/systemInfo")
-      .reply(200, systemInfoFixtures.showingNeither);
-  };
-
-  const setupAdminUser = () => {
-    axiosMock.reset();
-    axiosMock.resetHistory();
-    axiosMock
-      .onGet("/api/currentUser")
-      .reply(200, apiCurrentUserFixtures.adminUser);
-    axiosMock
-      .onGet("/api/systemInfo")
-      .reply(200, systemInfoFixtures.showingNeither);
-  };
-
-  const queryClient = new QueryClient();
-
-  test("Renders with Create Button for admin user", async () => {
-    setupAdminUser();
-    axiosMock.onGet("/api/restaurants/all").reply(200, []);
-
-    render(
-      <QueryClientProvider client={queryClient}>
-        <MemoryRouter>
-          <RestaurantIndexPage />
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-
-    await waitFor(() => {
-      expect(screen.getByText(/Create Restaurant/)).toBeInTheDocument();
-    });
-    const button = screen.getByText(/Create Restaurant/);
-    expect(button).toHaveAttribute("href", "/restaurants/create");
-    expect(button).toHaveAttribute("style", "float: right;");
-  });
-
-  test("renders three restaurants correctly for regular user", async () => {
-    setupUserOnly();
-    axiosMock
-      .onGet("/api/restaurants/all")
-      .reply(200, restaurantFixtures.threeRestaurants);
-
-    render(
-      <QueryClientProvider client={queryClient}>
-        <MemoryRouter>
-          <RestaurantIndexPage />
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-
-    await waitFor(() => {
-      expect(
-        screen.getByTestId(`${testId}-cell-row-0-col-id`),
-      ).toHaveTextContent("2");
-    });
-    expect(screen.getByTestId(`${testId}-cell-row-1-col-id`)).toHaveTextContent(
-      "3",
-    );
-    expect(screen.getByTestId(`${testId}-cell-row-2-col-id`)).toHaveTextContent(
-      "4",
-    );
-
-    const createRestaurantButton = screen.queryByText("Create Restaurant");
-    expect(createRestaurantButton).not.toBeInTheDocument();
-
-    const name = screen.getByText("Freebirds");
-    expect(name).toBeInTheDocument();
-
-    const description = screen.getByText(
-      "Burrito joint, and iconic Isla Vista location",
-    );
-    expect(description).toBeInTheDocument();
-
-    // for non-admin users, details button is visible, but the edit and delete buttons should not be visible
-    expect(
-      screen.queryByTestId("RestaurantTable-cell-row-0-col-Delete-button"),
-    ).not.toBeInTheDocument();
-    expect(
-      screen.queryByTestId("RestaurantTable-cell-row-0-col-Edit-button"),
-    ).not.toBeInTheDocument();
-  });
-
-  test("renders empty table when backend unavailable, user only", async () => {
-    setupUserOnly();
-
-    axiosMock.onGet("/api/restaurants/all").timeout();
-
-    const restoreConsole = mockConsole();
-
-    render(
-      <QueryClientProvider client={queryClient}>
-        <MemoryRouter>
-          <RestaurantIndexPage />
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-
-    await waitFor(() => {
-      expect(axiosMock.history.get.length).toBeGreaterThanOrEqual(1);
-    });
-
-    const errorMessage = console.error.mock.calls[0][0];
-    expect(errorMessage).toMatch(
-      "Error communicating with backend via GET on /api/restaurants/all",
-    );
-    restoreConsole();
-  });
-
-  test("what happens when you click delete, admin", async () => {
-    setupAdminUser();
-
-    axiosMock
-      .onGet("/api/restaurants/all")
-      .reply(200, restaurantFixtures.threeRestaurants);
-    axiosMock
-      .onDelete("/api/restaurants")
-      .reply(200, "Restaurant with id 1 was deleted");
-
-    render(
-      <QueryClientProvider client={queryClient}>
-        <MemoryRouter>
-          <RestaurantIndexPage />
-        </MemoryRouter>
-      </QueryClientProvider>,
-    );
-
-    await waitFor(() => {
-      expect(
-        screen.getByTestId(`${testId}-cell-row-0-col-id`),
-      ).toBeInTheDocument();
-    });
-
-    expect(screen.getByTestId(`${testId}-cell-row-0-col-id`)).toHaveTextContent(
-      "2",
-    );
-
-    const deleteButton = screen.getByTestId(
-      `${testId}-cell-row-0-col-Delete-button`,
-    );
-    expect(deleteButton).toBeInTheDocument();
-
-    fireEvent.click(deleteButton);
-
-    await waitFor(() => {
-      expect(mockToast).toBeCalledWith("Restaurant with id 1 was deleted");
-    });
-
-    await waitFor(() => {
-      expect(axiosMock.history.delete.length).toBe(1);
-    });
-    expect(axiosMock.history.delete[0].url).toBe("/api/restaurants");
-    expect(axiosMock.history.delete[0].url).toBe("/api/restaurants");
-    expect(axiosMock.history.delete[0].params).toEqual({ id: 2 });
-  });
-});
